Implement missing schema functions
